EAMxx: move test types definitions for test-all-scream into separate module
Makes test_all_scream shorter to parse and separate concerns
